MODS download and display Tab
Added MODS display tab, in a similar way on how the marcxml one works,
to allow partners so see full MODS records and download it. This MODS
is modified on the fly adding some extra elements referencing the
parent object (parent to the datastream). This will be probably
iterated  further by @ChrisStanton.